Fix close errors, move logging to new file
