+++ githubapp/middleware.go
@@ -15,17 +15,12 @@
 package githubapp
 
 import (
-	"bytes"
 	"fmt"
-	"io"
 	"net/http"
-	"regexp"
 	"strconv"
-	"time"
 
 	"github.com/gregjones/httpcache"
 	"github.com/rcrowley/go-metrics"
-	"github.com/rs/zerolog"
 )
 
 const (
@@ -111,131 +106,6 @@ func bucketStatus(status int) string {
 		return MetricsKeyRequests5xx
 	}
 	return ""
-}
-
-// ClientLogging creates client middleware that logs request and response
-// information at the given level. If the request fails without creating a
-// response, it is logged with a status code of -1. The middleware uses a
-// logger from the request context.
-func ClientLogging(lvl zerolog.Level, opts ...ClientLoggingOption) ClientMiddleware {
-	var options clientLoggingOptions
-	for _, opt := range opts {
-		opt(&options)
-	}
-
-	return func(next http.RoundTripper) http.RoundTripper {
-		return roundTripperFunc(func(r *http.Request) (*http.Response, error) {
-			var reqBody []byte
-			if requestMatches(r, options.RequestBodyPatterns) {
-				var err error
-				if r, reqBody, err = mirrorRequestBody(r); err != nil {
-					return nil, err
-				}
-			}
-
-			start := time.Now()
-			res, err := next.RoundTrip(r)
-			elapsed := time.Now().Sub(start)
-
-			evt := zerolog.Ctx(r.Context()).
-				WithLevel(lvl).
-				Str("method", r.Method).
-				Str("path", r.URL.String()).
-				Dur("elapsed", elapsed)
-
-			if reqBody != nil {
-				evt.Bytes("request_body", reqBody)
-			}
-
-			if res != nil {
-				cached := res.Header.Get(httpcache.XFromCache) != ""
-				evt.Int("status", res.StatusCode).
-					Int64("size", res.ContentLength).
-					Bool("cached", cached)
-			} else {
-				evt.Int("status", -1).
-					Int64("size", -1).
-					Bool("cached", false)
-			}
-
-			evt.Msg("github_request")
-			return res, err
-		})
-	}
-}
-
-// ClientLoggingOption controls behavior of client request logs.
-type ClientLoggingOption func(*clientLoggingOptions)
-
-// LogRequestBody enables request body logging for requests to paths matching
-// any of the regular expressions in patterns. It panics if any of the patterns
-// is not a valid regular expression.
-func LogRequestBody(patterns ...string) ClientLoggingOption {
-	regexps := compileRegexps(patterns)
-	return func(opts *clientLoggingOptions) {
-		opts.RequestBodyPatterns = regexps
-	}
-}
-
-// LogResponseBody enables response body logging for requests to paths matching
-// any of the regular expressions in patterns. It panics if any of the patterns
-// is not a valid regular expression.
-func LogResponseBody(patterns ...string) ClientLoggingOption {
-	regexps := compileRegexps(patterns)
-	return func(opts *clientLoggingOptions) {
-		opts.ResponseBodyPatterns = regexps
-	}
-}
-
-func mirrorRequestBody(r *http.Request) (*http.Request, []byte, error) {
-	switch {
-	case r.Body == nil:
-		return r, []byte{}, nil
-
-	case r.GetBody != nil:
-		br, err := r.GetBody()
-		if err != nil {
-			return r, nil, err
-		}
-		defer br.Close() // TODO(bkeyes): handle this error?
-
-		body, err := io.ReadAll(br)
-		return r, body, err
-
-	default:
-		defer r.Body.Close() // TODO(bkeyes): handle this error?
-
-		body, err := io.ReadAll(r.Body)
-		if err != nil {
-			return r, nil, err
-		}
-
-		rCopy := r.Clone(r.Context())
-		rCopy.Body = io.NopCloser(bytes.NewReader(body))
-		return rCopy, body, nil
-	}
-}
-
-func compileRegexps(pats []string) []*regexp.Regexp {
-	regexps := make([]*regexp.Regexp, len(pats))
-	for i, p := range pats {
-		regexps[i] = regexp.MustCompile(p)
-	}
-	return regexps
-}
-
-func requestMatches(r *http.Request, pats []*regexp.Regexp) bool {
-	for _, pat := range pats {
-		if pat.MatchString(r.URL.Path) {
-			return true
-		}
-	}
-	return false
-}
-
-type clientLoggingOptions struct {
-	RequestBodyPatterns  []*regexp.Regexp
-	ResponseBodyPatterns []*regexp.Regexp
 }
 
 type roundTripperFunc func(*http.Request) (*http.Response, error)
